Updated Social Bar styles.
Added a user tooltip to social bar user entries.
Introduced a button to expand and contract the social bar.
Added an option in the preferences page to enable or disable the social bar.
Introduced a new hook (settingsUpdated), which gets called whenever GSX settings are updated.
Re-enabled the social bar module.